Podcast Player: Break components into their own files (#15442)
* extract TrackIcon

* extract TrackError

* extract Track
